ssc_requirements.txt
in-toto
cryptography
cyclonedx-bom